Improved reading and writing in libbsfile
git-svn-id: https://femdk.l5.pk.edu.pl/bstools/BSTools/trunk@88 4ce36bd9-3a78-4973-a166-986cbda2e52d

+++ include/pkvaria.h
@@ -232,7 +232,7 @@ void pkv_ReverseMatc ( int nrows, int rowlen, int pitch, char *data );
     (outpitch)*sizeof(float),(char*)indata,(char*)outdata)
 #define pkv_Movef(nrows,rowlen,pitch,shift,data) \
   pkv_Movec(nrows,(rowlen)*sizeof(float),(pitch)*sizeof(float), \
-    (shift)*sizeof(float),(char*)data)
+    (shift)*(int)sizeof(float),(char*)data)
 #define pkv_ReverseMatf(nrows,rowlen,pitch,data) \
   pkv_ReverseMatc ( nrows, (rowlen)*sizeof(float), (pitch)*sizeof(float), \
     (char*)data )
@@ -248,7 +248,7 @@ void pkv_ReverseMatc ( int nrows, int rowlen, int pitch, char *data );
     (outpitch)*sizeof(double),(char*)indata,(char*)outdata)
 #define pkv_Moved(nrows,rowlen,pitch,shift,data) \
   pkv_Movec(nrows,(rowlen)*sizeof(double),(pitch)*sizeof(double), \
-    (shift)*sizeof(double),(char*)data)
+    (shift)*(int)sizeof(double),(char*)data)
 #define pkv_ReverseMatd(nrows,rowlen,pitch,data) \
   pkv_ReverseMatc ( nrows, (rowlen)*sizeof(double), (pitch)*sizeof(double), \
     (char*)data )
